Add compiler warnings and fix all warnings in src/
- Enable -Wall -Wextra -Wconversion -Wshadow -Wdouble-promotion
- Fix float/double literal mixing (use 273.0f not 273.0)
- Fix int to uint16_t conversion in rpms calculation
- Fix Serial.read() int to char conversion
- Change egtTemperatureC from double to float for consistency

Zero warnings from src/ code with aggressive compiler flags.

+++ include/AppData.h
@@ -24,7 +24,7 @@ struct AppData {
   float fuelPressurekPa = 0.0f;
   float fuelTemperatureC = 0.0f;
   float engineBayTemperatureC = 0.0f;
-  double egtTemperatureC = 0.0f;
+  float egtTemperatureC = 0.0f;
 };
 
 #endif  // OSSM_APPDATA_H
